Updated data source. Moved code to appropriate places. Code cleanup.
